# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FitnessTrack Public API
  version: 1.0.0
  description: >-
    REST API for the FitnessTrack platform. All endpoints respond with JSON and
    leverage RFC 7807 problem-details for errors.
servers:
  - url: /api/v1
    description: Version 1 base path
security:
  - bearerAuth: []
...
paths:
...
  /workouts/sessions:
    get:
      tags: [Workouts]
      operationId: listWorkoutSessions
      summary: List workout sessions
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Session page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSessionPage'
    post:
      tags: [Workouts]
      operationId: createWorkoutSession
      summary: Create workout session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutSession'
      responses:
        '201':
          description: Created session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSession'
  /workouts/sessions/{sessionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Workouts]
      operationId: getWorkoutSession
      summary: Retrieve workout session
      responses:
        '200':
          description: Session document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSession'
        '404':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Workouts]
      operationId: updateWorkoutSession
      summary: Update workout session
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutSessionUpdate'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSession'
    delete:
      tags: [Workouts]
      operationId: deleteWorkoutSession
      summary: Delete workout session
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
  /workouts/set-logs:
    get:
      tags: [Workouts]
      operationId: listExerciseSetLogs
      summary: List exercise logs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Log page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseSetLogPage'
    post:
      tags: [Workouts]
      operationId: createExerciseSetLog
      summary: Create exercise log
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseSetLog'
      responses:
        '201':
          description: Created log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseSetLog'
  /workouts/set-logs/{logId}:
    parameters:
      - name: logId
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Workouts]
      operationId: getExerciseSetLog
      summary: Retrieve log
      responses:
        '200':
          description: Log document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseSetLog'
        '404':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Workouts]
      operationId: updateExerciseSetLog
      summary: Update log
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseSetLogUpdate'
      responses:
        '200':
          description: Updated log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseSetLog'
    delete:
      tags: [Workouts]
      operationId: deleteExerciseSetLog
      summary: Delete log
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
      description: Page number (1-indexed)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
      description: Page size
    Sort:
      name: sort
      in: query
      schema:
        type: string
      description: Comma separated sort fields, prefix with '-' for descending.
...
  responses:
    Problem:
      description: Problem details error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        errors:
          oneOf:
            - type: object
            - type: array
              items:
                type: object
    PageEnvelope:
      type: object
      properties:
        items:
          type: array
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      required: [items, page, limit, total]
...
    WorkoutSession:
      type: object
      properties:
        id:
          type: integer
        subject_id:
          type: integer
        workout_date:
          type: string
          format: date-time
        status:
          type: string
        routine_day_id:
          type: integer
          nullable: true
        cycle_id:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        perceived_fatigue:
          type: integer
          nullable: true
        bodyweight_kg:
          type: number
          format: float
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkoutSessionUpdate:
      type: object
      properties:
        workout_date:
          type: string
          format: date-time
        status:
          type: string
        routine_day_id:
          type: integer
        cycle_id:
          type: integer
        location:
          type: string
        perceived_fatigue:
          type: integer
        bodyweight_kg:
          type: number
          format: float
        notes:
          type: string
    WorkoutSessionPage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/WorkoutSession'
    ExerciseSetLog:
      type: object
      properties:
        id:
          type: integer
        subject_id:
          type: integer
        exercise_id:
          type: integer
        session_id:
          type: integer
          nullable: true
        planned_set_id:
          type: integer
          nullable: true
        performed_at:
          type: string
          format: date-time
        set_index:
          type: integer
        is_warmup:
          type: boolean
        to_failure:
          type: boolean
        actual_weight_kg:
          type: number
          format: float
          nullable: true
        actual_reps:
          type: integer
          nullable: true
        actual_rir:
          type: integer
          nullable: true
        actual_rpe:
          type: number
          format: float
          nullable: true
        actual_tempo:
          type: string
          nullable: true
        actual_rest_s:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExerciseSetLogUpdate:
      type: object
      properties:
        exercise_id:
          type: integer
        session_id:
          type: integer
        planned_set_id:
          type: integer
        performed_at:
          type: string
          format: date-time
        set_index:
          type: integer
        is_warmup:
          type: boolean
        to_failure:
          type: boolean
        actual_weight_kg:
          type: number
          format: float
        actual_reps:
          type: integer
        actual_rir:
          type: integer
        actual_rpe:
          type: number
          format: float
        actual_tempo:
          type: string
        actual_rest_s:
          type: integer
        notes:
          type: string
    ExerciseSetLogPage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ExerciseSetLog'